decouple PriceType
